requirements.txt
python-dotenv
langgraph
langchain
langchain_community
langchain_huggingface
langchain_openai
numpy
pandas
typing_extensions
ipython
ipykernel